feat: Migrate to appDir - collections & 404 pages
