+++ demo-app/src/app/not-found.tsx
@@ -1,0 +1,9 @@
+import notFound from "design-system/not-found.module.css";
+
+export default function Page404() {
+  return (
+    <section className={notFound.section}>
+      <h1>404 - Page Not Found</h1>
+    </section>
+  );
+}
