--- demo-app/src/components/Custom404.ts
@@ -1,7 +1,0 @@
-import styled from "styled-components";
-
-export const Section404 = styled.section`
-  height: 100%;
-  display: grid;
-  place-content: center;
-`;
