+++ demo-app/src/design-system/button-group.module.css
@@ -1,0 +1,11 @@
+.btnGroup {
+  --btn-gap: 0.5rem;
+  margin: 0 auto;
+  text-align: center;
+
+  margin-bottom: 2rem;
+
+  display: flex;
+  flex-wrap: wrap;
+  gap: var(--btn-gap, 0.5rem);
+}
