+++ demo-app/src/design-system/not-found.module.css
@@ -1,0 +1,5 @@
+.section {
+  height: 100%;
+  display: grid;
+  place-content: center;
+}
