--- demo-app/src/pages/404.tsx
@@ -1,9 +1,0 @@
-import { Section404 } from "components/Custom404";
-
-export default function Page404() {
-  return (
-    <Section404>
-      <h1>404 - Page Not Found</h1>
-    </Section404>
-  );
-}
